requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.10
iniconfig==2.0.0
packaging==24.1
playwright==1.47.0
pluggy==1.5.0
pyee==12.0.0
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.5.2
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
